refactor(Advanced.vue, User.vue): use non-null assertion operator for boolean input fields
The non-null assertion operator is used to indicate that a value will not be null or undefined. This operator is used in the input fields of type boolean in the Advanced.vue and User.vue files to ensure that the value of the input field is not null or undefined.